feat(agent-runner): complete merged PR sync
