+++ scripts/tests/agent-runner-lifecycle.test.mjs
@@ -11,19 +11,6 @@ import {
   commandRunFieldUpdate,
 } from "../lib/agent-runner-execution.mjs"
 import {
-  evaluatePullRequestCompletion,
-  evaluatePullRequestGate,
-  isRemoteReference,
-  pullRequestBody,
-  pullRequestCompletionFieldValues,
-  pullRequestCreateArgs,
-  pullRequestFieldValues,
-  pullRequestNumberFromUrl,
-  pullRequestSyncFieldValues,
-  pullRequestTitle,
-  summarizeChecks,
-} from "../lib/agent-runner-pr.mjs"
-import {
   buildExecutionPlan,
   canCleanupAgentState,
   cleanupFieldValues,
@@ -256,244 +243,5 @@ describe("agent runner lifecycle helpers", () => {
       console.error = original.error
       process.exit = original.exit
     }
-  })
-
-  it("builds PR metadata from issue and evidence context", () => {
-    const item = workItem()
-    item.issue.title = "[Task] Publish runner PR metadata"
-    item.fields = {
-      Branch: "task/579-test-agent-project-intake-workflow",
-    }
-
-    assert.equal(pullRequestTitle(item), "Publish runner PR metadata")
-    assert.deepEqual(
-      pullRequestFieldValues({
-        date: new Date("2026-05-10T12:34:56.000Z"),
-        prUrl: "https://github.com/voyantjs/voyant/pull/601",
-      }),
-      {
-        PR: "https://github.com/voyantjs/voyant/pull/601",
-        "Last Heartbeat": "2026-05-10",
-      },
-    )
-
-    const body = pullRequestBody(item, {
-      evidenceBody: "# Evidence\n\nAll checks passed.",
-      evidenceReference: "docs/agent-evidence/active/579-test.md",
-      repository: "voyantjs/voyant",
-    })
-
-    assert.match(body, /Implements #579/)
-    assert.match(body, /<summary>docs\/agent-evidence\/active\/579-test.md<\/summary>/)
-    assert.match(body, /All checks passed/)
-    assert.match(body, /Verification lane: verify:fast/)
-    assert.doesNotMatch(body, /codex|claude|generated by/i)
-  })
-
-  it("formats remote evidence as a link in PR bodies", () => {
-    const body = pullRequestBody(workItem(), {
-      evidenceBody: "",
-      evidenceReference: "https://github.com/voyantjs/voyant/issues/579#issuecomment-1",
-      repository: "voyantjs/voyant",
-    })
-
-    assert.equal(isRemoteReference("https://github.com/voyantjs/voyant/issues/579"), true)
-    assert.equal(isRemoteReference("docs/agent-evidence/active/579-test.md"), false)
-    assert.match(body, /- https:\/\/github\.com\/voyantjs\/voyant\/issues\/579#issuecomment-1/)
-    assert.doesNotMatch(body, /<details>/)
-  })
-
-  it("opens same-repository PRs with an unqualified head branch", () => {
-    const args = pullRequestCreateArgs({
-      base: "main",
-      body: "PR body",
-      branch: "task/579-test-agent-project-intake-workflow",
-      draft: true,
-      title: "Test agent project intake workflow",
-    })
-
-    assert.deepEqual(args, [
-      "pr",
-      "create",
-      "--base",
-      "main",
-      "--head",
-      "task/579-test-agent-project-intake-workflow",
-      "--title",
-      "Test agent project intake workflow",
-      "--body",
-      "PR body",
-      "--draft",
-    ])
-    assert.equal(args.includes("voyantjs:task/579-test-agent-project-intake-workflow"), false)
-  })
-
-  it("parses PR numbers from GitHub pull request URLs", () => {
-    assert.equal(pullRequestNumberFromUrl("https://github.com/voyantjs/voyant/pull/603"), 603)
-    assert.equal(
-      pullRequestNumberFromUrl("https://github.com/voyantjs/voyant/issues/603"),
-      undefined,
-    )
-  })
-
-  it("summarizes PR checks into successful, pending, and failed buckets", () => {
-    assert.deepEqual(
-      summarizeChecks([
-        { name: "build", status: "COMPLETED", conclusion: "SUCCESS" },
-        { name: "checks", status: "IN_PROGRESS", conclusion: "" },
-        { name: "test", status: "COMPLETED", conclusion: "FAILURE" },
-      ]),
-      {
-        failed: ["test"],
-        pending: ["checks"],
-        successful: ["build"],
-      },
-    )
-  })
-
-  it("moves failing PR checks to CI repair", () => {
-    assert.deepEqual(
-      evaluatePullRequestGate({
-        state: "OPEN",
-        isDraft: false,
-        reviewDecision: "",
-        statusCheckRollup: [{ name: "checks", status: "COMPLETED", conclusion: "FAILURE" }],
-      }),
-      {
-        agentState: "CI Repair",
-        blockedBy: "Failing checks: checks",
-        mergeReady: false,
-        reason: "checks failing",
-      },
-    )
-  })
-
-  it("moves requested changes to changes requested before check state", () => {
-    assert.deepEqual(
-      evaluatePullRequestGate({
-        state: "OPEN",
-        isDraft: false,
-        reviewDecision: "CHANGES_REQUESTED",
-        statusCheckRollup: [{ name: "checks", status: "COMPLETED", conclusion: "SUCCESS" }],
-      }),
-      {
-        agentState: "Changes Requested",
-        blockedBy: "PR changes requested",
-        mergeReady: false,
-        reason: "review changes requested",
-      },
-    )
-  })
-
-  it("keeps draft or pending-check PRs in human review", () => {
-    assert.equal(
-      evaluatePullRequestGate({
-        state: "OPEN",
-        isDraft: true,
-        reviewDecision: "",
-        statusCheckRollup: [{ name: "checks", status: "COMPLETED", conclusion: "SUCCESS" }],
-      }).agentState,
-      "Human Review",
-    )
-
-    assert.deepEqual(
-      evaluatePullRequestGate({
-        state: "OPEN",
-        isDraft: false,
-        reviewDecision: "",
-        statusCheckRollup: [{ name: "checks", status: "QUEUED", conclusion: "" }],
-      }),
-      {
-        agentState: "Human Review",
-        blockedBy: "Pending checks: checks",
-        mergeReady: false,
-        reason: "checks pending",
-      },
-    )
-  })
-
-  it("keeps review-required PRs in human review even with passing checks", () => {
-    assert.deepEqual(
-      evaluatePullRequestGate({
-        state: "OPEN",
-        isDraft: false,
-        reviewDecision: "REVIEW_REQUIRED",
-        statusCheckRollup: [{ name: "checks", status: "COMPLETED", conclusion: "SUCCESS" }],
-      }),
-      {
-        agentState: "Human Review",
-        blockedBy: "PR review decision: REVIEW_REQUIRED",
-        mergeReady: false,
-        reason: "review required",
-      },
-    )
-  })
-
-  it("moves passing non-draft PRs to merge ready", () => {
-    const pr = {
-      url: "https://github.com/voyantjs/voyant/pull/603",
-      state: "OPEN",
-      isDraft: false,
-      reviewDecision: "",
-      statusCheckRollup: [{ name: "checks", status: "COMPLETED", conclusion: "SUCCESS" }],
-    }
-    const result = evaluatePullRequestGate(pr)
-
-    assert.deepEqual(result, {
-      agentState: "Merge Ready",
-      blockedBy: null,
-      mergeReady: true,
-      reason: "PR is ready for maintainer merge",
-    })
-    assert.deepEqual(
-      pullRequestSyncFieldValues({
-        date: new Date("2026-05-10T12:34:56.000Z"),
-        pr,
-        result,
-      }),
-      {
-        "Agent State": "Merge Ready",
-        PR: "https://github.com/voyantjs/voyant/pull/603",
-        "Last Heartbeat": "2026-05-10",
-      },
-    )
-  })
-
-  it("marks merged PRs complete", () => {
-    const pr = {
-      url: "https://github.com/voyantjs/voyant/pull/605",
-      state: "MERGED",
-    }
-    const result = evaluatePullRequestCompletion(pr)
-
-    assert.deepEqual(result, {
-      complete: true,
-      reason: "PR is merged",
-    })
-    assert.deepEqual(
-      pullRequestCompletionFieldValues({
-        date: new Date("2026-05-10T12:34:56.000Z"),
-        pr,
-      }),
-      {
-        Status: "Done",
-        "Agent State": "Done",
-        PR: "https://github.com/voyantjs/voyant/pull/605",
-        "Last Heartbeat": "2026-05-10",
-      },
-    )
-  })
-
-  it("refuses to complete unmerged PRs", () => {
-    assert.deepEqual(
-      evaluatePullRequestCompletion({
-        url: "https://github.com/voyantjs/voyant/pull/605",
-        state: "OPEN",
-      }),
-      {
-        complete: false,
-        reason: "PR is OPEN",
-      },
-    )
   })
 })
